backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.2.11
langchain-core==0.2.23
langchain-google-genai==1.0.7
langchain-community==0.2.10
pytest==7.4.3
pytest-asyncio==0.23.2
